Bluetooth: btusb: Add reset_hci handler
After users update the bluetooth firmware, they want the driver to
load the new firmware, with the current driver, users could not make
sure the driver will load the new firmware, on some laptops, they have
to run poweroff and cold boot the machine, on some desktops, they have
to wait for a couple of seconds between poweroff and cold boot, it is
very annoying to users.

Here adding a reset_hci handler, for hci on the USB bus, we provide
a generic method, this could make the device reset over the USB bus,
it will introduce calling disconnect() and probe(), but this doesn't
guarantee the hci enters bootloader mode and load the new firmware.
For Intel_new and Intel_newgen hcis, there is a method to make them
enter bootloader mode and reset over the USB bus, so we use specific
method to overwrite the generic method for those hcis.

If we have more specific methods to make some hcis enter bootloader
mode in future, we could overwrite the generic method for them like
Intel hcis.

Signed-off-by: Hui Wang <hui.wang@canonical.com>