Bluetooth: NXP: Add protocol support for NXP Bluetooth chipsets
This adds a driver based on serdev driver for the NXP BT serial
protocol based on running H:4, which can enable the built-in
Bluetooth device inside a generic NXP BT chip.

This driver has Power Save feature that will put the chip into
sleep state whenever there is no activity for 2000ms, and will
be woken up when any activity is to be initiated.

This driver enables the power save feature by default by sending
the vendor specific commands to the chip during setup.

During setup, the driver is capable of validating correct chip
is attached to the host based on the compatibility parameter
from DT and chip's unique bootloader signature, and download
firmware.

Signed-off-by: Neeraj Sanjay Kale <neeraj.sanjaykale@nxp.com>