Bluetooth: ISO: Set CIS bit only for devices with CIS support
As mentioned in the Bluetooth Specification the full name of this bit is
"Connected Isochronous Stream (Host Support)", thus it should be set only
for CIS capable devices. Currently, it is set for any device that has CIS
or BIS support. In reality, devices that support BIS may not allow that
bit to be set and so, the HCI bring up fails for them.

This commit fixes this by only setting the bit for CIS capable devices.

Signed-off-by: Vlad Pruteanu <vlad.pruteanu@nxp.com>