Bluetooth: add experimental BT_NO_ERRQUEUE_POLL socket option
Add experimental feature that enables a SOL_BLUETOOTH socket option to
disable POLLERR on non-empty socket error queue.  Add corresponding MGMT
experimental feature UUID for it.

This is intended to allow applications disable the POLLERR wakeups due
to TX timestamps, and allow upper layers of the stack to enable and
consume TX timestamps, without needing to complicate the lower layer
POLLERR handling which is only interested in real socket errors.

It is a socket option disabled by default, as it enables a deviation
from the common net/ TX timestamping API.

Signed-off-by: Pauli Virtanen <pav@iki.fi>