Bluetooth: add experimental BT_POLL_ERRQUEUE socket option
Add experimental feature that enables a SOL_BLUETOOTH socket option to
disable POLLERR on non-empty socket error queue.  Add corresponding MGMT
experimental feature UUID for it.

This is intended to allow applications disable the POLLERR wakeups due
to TX timestamps, and allow upper layers of the stack to enable and
consume TX timestamps, without needing to complicate the lower layer
POLLERR handling which is only interested in real socket errors.

It is a socket option as it enables a deviation from the common net/
errqueue API.

Signed-off-by: Pauli Virtanen <pav@iki.fi>
Signed-off-by: Luiz Augusto von Dentz <luiz.von.dentz@intel.com>