wifi: wilc1000: add a read-modify-write API for registers accesses
Many places in wilc driver need to modify a single bit in a register,
and then perform the following sequence: read the register, set/reset
the needed bit, write the register. This sequence is performed in
multiple places in the driver, with a varying amount of checks around
possible errors.

Replace all those sequences with a new hif_rmw_reg helper. Make sure to
perform the write only if needed.

Signed-off-by: Alexis Lothoré <alexis.lothore@bootlin.com>